+++ aegis_control/launch/ft_sensor_driver.launch.py
@@ -6,7 +6,7 @@ import sys
 
 run_path = os.path.dirname(os.path.realpath(__file__))
 sys.path.append(run_path)
-from include.utils import controllers_spawner
+from include.utils import controllers_spawner  # noqa E402
 
 
 def generate_launch_description():