+++ aegis_control/launch/start_drivers.launch.py
@@ -12,7 +12,6 @@ from launch_ros.substitutions import FindPackageShare
 
 
 def generate_launch_description() -> LaunchDescription:
-
     launch_args = {
         "tf_prefix": LaunchConfiguration("tf_prefix", default=""),
         "mock_hardware": LaunchConfiguration("mock_hardware", default="false"),
@@ -55,6 +54,7 @@ def generate_launch_description() -> LaunchDescription:
                 ]
             )
         ),
+        launch_arguments=launch_args.items(),
     )
 
     gripper_driver = IncludeLaunchDescription(