+++ aegis_description/launch/aegis_preview.launch.py
@@ -1,5 +1,4 @@
 from launch import LaunchDescription
-from launch_ros.actions import Node
 from launch_ros.substitutions import FindPackageShare
 from launch.actions import IncludeLaunchDescription
 from launch.substitutions import PathJoinSubstitution